+++ age/CHANGELOG.md
@@ -12,8 +12,13 @@ to 1.0.0 are beta releases.
 ### Added
 - `age::armor::ArmoredReader`, which can be wrapped around an input to handle
   a potentially-armored age file.
+- `age::armor::ArmoredWriter`, which can be wrapped around an output to
+  optionally apply the armored age format.
 
 ### Changed
+- `age::Encryptor::wrap_output` now only generates the non-malleable binary age
+  format. Use `encryptor.wrap_output(age::armor::ArmoredWriter(output, format))`
+  to optionally generate armored age files.
 - `age::Decryptor` now only decrypts the non-malleable binary age format. Use
   `age::Decryptor::new(age::armor::ArmoredReader::new(input))` to handle age
   files that are potentially armored.